Update README.md
The GoatPro app was designed for handling, visualizing and making small biomechanically accurate adjustments to geometric morphometric data on lateral images of goatfish specimens. Specimen photos are often variable in the fin angles, the tail spread, the barbel curvature and angle, and the degree to which the lower jaw was open. This variability is an artifact of preservation due to the fin or barbel positions when the photograph was taken. GoatPro was developed to adjust each of these elements to a standard position. All standard positions are readily attained with a fresh fish and are within the behavioral and biomechanical range of motion for all species. The app rotates each fish to the horizontal axis and enables a set of batch processing (all specimens) and custom functions (individual specimen) to biomechanically rotate peripheral anatomical elements. Batch functions straighten the barbels and set them to a posteroventral 45-degree angle, rotate all dorsal, anal, and pelvic fin leading edges to a posteroventral 45-degree angle, and all trailing edges to a 10-degree angle to the horizontal. Individuals with caudle peduncle or caudal fin slightly displaced dorsally or ventrally may be straightened while maintaining all other coordinate positions. The caudal fin may be spread by rotating dorsal and ventral lobes independently, to attain fin edge angles of 20-degrees to the horizontal. Finally, for specimens with the mouth open, the lower jaw can be rotated dorsally to close the mouth, through biomechanically computed jaw rotations, with the barbel and barbel base tracking the jaw as it closed. The app will export coordinate sets for all specimens with their adjusted positions to a file used for morphometric analysis. The rotated coordinate file output is in an x x x x y y y y  space delimited text file. The app also computes a large set of body and fin metrics for goatfish specimens, including various lengths, fin areas, fin spans, and aspect ratios.